jokes_project/requirements.txt
asgiref==3.6.0
Django==4.0
mssql-django==1.2
pyodbc==4.0.35
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
